Split tests to multiple files
- Split tests to multiple files for easier management and better performance in editor.
- Created a utility module for VCR configuration to streamline cassette management.
