+++ tests/test_prices_not_available.py
@@ -1,0 +1,11 @@
+import unittest
+from nordpool.elspot import Prices
+from ._utils import vcr
+
+
+class ElspotTestCase(unittest.TestCase):
+    def test_prices_not_available(self):
+        with vcr.use_cassette("prices_not_available.yaml"):
+            elspot = Prices()
+            prices = elspot.fetch(elspot.HOURLY, end_date="2024-10-17", areas=["FI"])
+            self.assertIsNone(prices)
